+++ aleksis/core/util/core_helpers.py
@@ -208,11 +208,23 @@ def custom_information_processor(request: HttpRequest) -> dict:
         regrouped_pwa_icons.setdefault(pwa_icon.rel, {})
         regrouped_pwa_icons[pwa_icon.rel][pwa_icon.size] = pwa_icon
 
-    return {
+    context = {
         "FOOTER_MENU": CustomMenu.get_default("footer"),
         "ADMINS": settings.ADMINS,
         "PWA_ICONS": regrouped_pwa_icons,
+        "SENTRY_ENABLED": settings.SENTRY_ENABLED,
     }
+
+    if settings.SENTRY_ENABLED:
+        context["SENTRY_SETTINGS"] = settings.SENTRY_SETTINGS
+
+        import sentry_sdk
+
+        span = sentry_sdk.Hub.current.scope.span
+        if span is not None:
+            context["SENTRY_TRACE_ID"] = span.to_traceparent()
+
+    return context
 
 
 def now_tomorrow() -> datetime: