+++ CHANGELOG.rst
@@ -13,12 +13,16 @@ Added
 ~~~~~
 
 * Provide an ``ExtensiblePolymorphicModel`` to support the features of extensible models for polymorphic models and vice-versa.
+* Implement optional Sentry integration for error and performance tracing.
+* Support trusted OAuth applications that leave out the authorisation screen.
 * Add birthplace to Person model.
 
 Changed
 ~~~~~~~
 
 * Replace dev.sh helper script with tox environments.
+* OAuth Grant Flows are now configured system-wide instead of per app.
+* Refactor OAuth2 application management views.
 
 `2.0`_ - 2021-10-29
 -------------------